series

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -18,13 +18,13 @@ int main(void)
  */
 void print_f(int limit)
 {
-	 unsigned long long int i, result, a, b;
+	 unsigned long int i, result, a, b;
 
 	a = 1;
 	b = 2;
 	for (i = 0; i < limit; i++)
 	{
-		printf("%llu", a);
+		printf("%lu", a);
 		if (i != (limit - 1))
 			printf(", ");
 		result = a + b;
